--- a/lib_eio/unix/include/eio_unix_stubs.h
+++ b/lib_eio/unix/include/eio_unix_stubs.h
@@ -1,0 +1,7 @@
+#include <caml/mlvalues.h>
+
+/* A function to convert directory entry kinds into
+ * Unix.file_kind options.
+ */
+CAMLprim value eio_unix_file_type_of_dtype(int d_type);
+