--- a/lib_eio_posix/fs.ml
+++ b/lib_eio_posix/fs.ml
@@ -87,7 +87,7 @@ end = struct
     |> Array.to_list
 
   let with_dir_entries t path fn =
-    Err.run (Low_level.with_dir_entries t.fd path) fn
+    Low_level.with_dir_entries t.fd path fn
 
   let read_link t path =
     Err.run (Low_level.read_link t.fd) path