--- a/lib_eio_posix/primitives.h
+++ b/lib_eio_posix/primitives.h
@@ -11,6 +11,7 @@ CAMLprim value caml_eio_posix_preadv(value, value, value);
 CAMLprim value caml_eio_posix_pwritev(value, value, value);
 CAMLprim value caml_eio_posix_openat(value, value, value, value);
 CAMLprim value caml_eio_posix_fdopendir(value);
+CAMLprim value caml_eio_posix_readdir(value);
 CAMLprim value caml_eio_posix_mkdirat(value, value, value);
 CAMLprim value caml_eio_posix_unlinkat(value, value, value);
 CAMLprim value caml_eio_posix_renameat(value, value, value, value);