@patricoferris 2025-11-03
Thanks @avsm !

Note that the current implementation does _not_ descend into directories (see the recursive call to `walk` in the benchmark above). Happy to change the behaviour to recurse by default.